requirements.txt
requests
six
kafka-python==2.0.0
kafka-python-ng
fastapi
uvicorn